refactor: change location for Navigation and UnorderedList
