+++ nav-bar-sticky/src/App.tsx
@@ -1,6 +1,6 @@
-import { Navigation } from './components/nav-bar/Navigation';
+import { Navigation } from './components/Navigation';
 import { useAppSelector } from './app/hooks';
-import { displayPage } from './components/nav-bar/pageSlice';
+import { displayPage } from './components/un-list/pageSlice';
 import Container from './components/Container';
 import Home from './components/Home';
 import About from './components/About';
