+++ nav-bar-sticky/src/components/un-list/UnorderedList.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { useAppDispatch } from '../../app/hooks';
-import { changer } from '../nav-bar/pageSlice';
+import { changer } from './pageSlice';
 import Item from './Item';
 import Link from '../Link';
